% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_traindat.R
\name{form_traindat}
\alias{form_traindat}
\title{Format training data for plotting}
\usage{
form_traindat(trndat)
}
\arguments{
\item{trndat}{output from \code{\link{read_traindat}}}
}
\value{
data frame in long format
}
\description{
Format training data for plotting
}
\examples{
form_traindat(trndat)
}
