% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_trndat.R
\name{form_trndat}
\alias{form_trndat}
\title{Format training data from file for plotting}
\usage{
form_trndat(trndat)
}
\arguments{
\item{trndat}{output from \code{\link{read_trndat}}}
}
\value{
data frame in long format
}
\description{
Format training data from file for plotting
}
\examples{
form_trndat(trndat)
}
