% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_trnjsn.R
\name{form_trnjsn}
\alias{form_trnjsn}
\title{Format transect data from Water Atlas for plotting}
\usage{
form_trnjsn(trnjsn)
}
\arguments{
\item{trnjsn}{output from \code{\link{read_trnjsn}}}
}
\value{
data frame in long format
}
\description{
Format transect data from Water Atlas for plotting
}
\examples{
trnjsn <- read_trnjsn()
form_trnjsn(trnjsn)
}
