% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_traindat.R
\name{read_traindat}
\alias{read_traindat}
\title{Import training data}
\usage{
read_traindat(pth, sheet = "All data")
}
\arguments{
\item{pth}{chr string for location of file}

\item{sheet}{chr string of sheet to import}
}
\value{
data frame
}
\description{
Import training data
}
\examples{
\dontrun{
pth <- 'inst/extdata/Class_Data_20190627.xlsx'
read_traindat(pth)
}
}
