% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trndat.R
\name{read_trndat}
\alias{read_trndat}
\title{Import training data from file}
\usage{
read_trndat(pth, sheet = "All data")
}
\arguments{
\item{pth}{chr string for location of file}

\item{sheet}{chr string of sheet to import}
}
\value{
data frame
}
\description{
Import training data from file
}
\examples{
\dontrun{
pth <- 'inst/extdata/Class_Data_20190627.xlsx'
read_trndat(pth)
}
}
