% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trnjsn.R
\name{read_trnjsn}
\alias{read_trnjsn}
\title{Import JSON training data from Water Atlas}
\usage{
read_trnjsn(
  url = "http://dev.seagrass.wateratlas.usf.edu/api/assessments/training"
)
}
\arguments{
\item{url}{chr string of API endpoint}
}
\value{
data frame
}
\description{
Import JSON training data from Water Atlas
}
\examples{
trnjsn <- read_trnjsn()
}
