% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_compplot.R
\name{show_compplot}
\alias{show_compplot}
\title{Make a bar plot for training group comparisons}
\usage{
show_compplot(
  dat,
  site = c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10"),
  species = c("Halodule", "Thalassia"),
  varplo = c("Abundance", "Blade Length", "Short Shoot Density"),
  base_size = 18
)
}
\arguments{
\item{dat}{data frame returned by \code{\link{form_traindat}}}

\item{site}{numeric value indicating site results to plot}

\item{species}{chr string indicating which species to plot}

\item{varplo}{chr string indicating which variable to plot}

\item{base_size}{numeric indicating text scaling size for plot}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Make a bar plot for training group comparisons
}
\examples{
dat <- form_traindat(trndat)
show_compplot(dat)
}
