% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trndat.R
\docType{data}
\name{trndat}
\alias{trndat}
\title{Imported training data}
\format{
A data frame
}
\usage{
trndat
}
\description{
Imported training data
}
\examples{
\dontrun{
pth <- 'inst/extdata/Class_Data_20190627.xlsx'
trndat <- read_traindat(pth)
save(trndat, file = here::here('data', 'trndat.RData'), compress = 'xz')
}
}
\keyword{datasets}
