% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnlns.R
\docType{data}
\name{trnlns}
\alias{trnlns}
\title{Seagrass transect locations}
\format{A \code{sf} LINESTRING object}
\usage{
trnlns
}
\description{
Seagrass transect locations
}
\examples{
\dontrun{
library(sf)

trnlns <- st_read('T:/05_GIS/SEAGRASS_TRANSECTS/transect_routes.shp') \%>\% 
   st_transform(crs = 4326)

save(trnlns, file = 'data/trnlns.RData', compress = 'xz')
}
}
\keyword{datasets}
