% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnpts.R
\docType{data}
\name{trnpts}
\alias{trnpts}
\title{Seagrass transect starting locations}
\format{A \code{sf} POINT object}
\usage{
trnpts
}
\description{
Seagrass transect starting locations
}
\examples{
\dontrun{
library(sf)

trnpts <- st_read('T:/05_GIS/SEAGRASS_TRANSECTS/TransectBasics2019.shp') \%>\% 
   st_transform(crs = 4326)

save(trnpts, file = 'data/trnpts.RData', compress = 'xz')
}
}
\keyword{datasets}
